+++ JumpProto/RectCoordBuffer.m
@@ -185,7 +185,7 @@
 
     glLoadIdentity();
     
-    const size_t numIndicesToDraw = 6 * m_currentFrame;  // 2 triangles per frame
+    const int numIndicesToDraw = 6 * m_currentFrame;  // 2 triangles per frame
     glDrawArrays( GL_TRIANGLES, 0, numIndicesToDraw );
     m_currentFrame = 0;
 }
